+++ objects/State.lua
@@ -473,6 +473,9 @@ function State:serialize()
     -- Castle Kingdoms 2027 v3.11.913: Save Royal Market auto-sell state
     local RMI = require("objects.Economy.RoyalMarketIntegration")
     data.royalMarket = RMI.serialize()
+    -- Castle Kingdoms 2027 v3.11.914: Save DynamicMarket state (prices, inflation, events)
+    local DynamicMarket = require("objects.Economy.DynamicMarketSystem")
+    data.dynamicMarket = DynamicMarket.serialize()
     return data, metadata
 end
 
@@ -563,6 +566,11 @@ function State:deserialize(load)
     if load.royalMarket then
         local RMI = require("objects.Economy.RoyalMarketIntegration")
         RMI.deserialize(load.royalMarket)
+    end
+    -- Castle Kingdoms 2027 v3.11.914: Load DynamicMarket state (prices, inflation, events)
+    if load.dynamicMarket then
+        local DynamicMarket = require("objects.Economy.DynamicMarketSystem")
+        DynamicMarket.deserialize(load.dynamicMarket)
     end
     self.scaleX = load.scaleX
     self.viewXview = load.viewXview